+++ napari_spatialdata/__init__.py
@@ -1,7 +1,7 @@
 # from napari_spatial_anndata._reader import napari_get_reader
 # from napari_spatial_anndata._widget import ExampleQWidget, example_magic_widget
 # from napari_spatial_anndata._writer import write_multiple, write_single_image
-from napari_spatial_anndata.interactive import Interactive
+from napari_spatialdata.interactive import Interactive
 
 # from napari_spatial_anndata._sample_data import make_sample_data
