+++ napari_spatialdata/_constants/_constants.py
@@ -1,7 +1,7 @@
 """Constants that user deals with."""
 from enum import unique
 
-from napari_spatial_anndata._constants._utils import ModeEnum
+from napari_spatialdata._constants._utils import ModeEnum
 
 
 @unique
