+++ napari_spatialdata/_constants/_pkg_constants.py
