+++ napari_spatialdata/_constants/_utils.py
