+++ napari_spatialdata/napari.yaml
